Complete story for userroles and authorization

+++ app/controllers/wikis_controller.rb
@@ -1,11 +1,14 @@
 class WikisController < ApplicationController
   
+  before_action :require_login
+  
   def index
-    @wikis = Wiki.all
+    @wikis = policy_scope(Wiki)
   end
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
@@ -14,8 +17,9 @@ class WikisController < ApplicationController
   
   def create
     @wiki = Wiki.new(wiki_params)
-    @wiki.user_id = params[:wiki][:user]||= current_user.id
     
+    authorize @wiki
+    @wiki.user_id = params[:wiki][:user]||= current_user.id
    
     if @wiki.save
       flash[:notice] = "Your wiki has been successfully created.Yay!"
@@ -28,25 +32,28 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
   
   def update
     @wiki = Wiki.find(params[:id])
     
+    authorize @wiki
     @wiki.assign_attributes(wiki_params) 
     
     if @wiki.save
       flash[notice] = "Wiki has been successfully updated!"
       redirect_to [@wiki]
     else
-      flash.now[:alert] = "Bummer!There was an error trying to save the wiki changes..Please try again later"
+      flash.now[:alert] = "Bummer!There was an error trying to save the wiki changes.Please try again later"
       render :edit
     end
   end  
   
   def destroy
     @wiki = Wiki.find(params[:id])
- 
+    
+    authorize @wiki
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
       redirect_to action: :index
@@ -61,7 +68,5 @@ class WikisController < ApplicationController
   def wiki_params
     params.require(:wiki).permit(:title, :body)
   end
-  
-  
   
 end
